+++ src/leotools/__init__.py
@@ -8,7 +8,7 @@ leotools
 - preproc: Satellite image preprocessing tools.
 """
 
-__version__ = "0.8.0"
+__version__ = "0.9.0"
 
 __author__ = "Rómeó Komolafe"